added caption to image on home page
resolved:
Please could you add the image and video titles as in the brochure - that is the light yellow small text? perhaps this does not work in a responsive version? I got a lot of feedback from the team that this is not so closely aligned to the brochure - is there anything we can do to make this more like that?

and update to close button on video overlay, and